requirements-worker.txt
# Storage Worker Requirements
kafka-python>=2.0.2
boto3>=1.34.0
pyarrow>=14.0.2
pandas>=2.1.4
prometheus-client>=0.19.0
pyyaml>=6.0.1
aiohttp>=3.9.1